// api/consolidated/models/StringObjectDictionaryListOperationResult.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { EntityValidationError } from './EntityValidationError';

export type StringObjectDictionaryListOperationResult = {
    outcome?: StringObjectDictionaryListOperationResult.outcome;
    errors?: Array<EntityValidationError> | null;
    data?: Array<Record<string, any>> | null;
    correlationId?: string | null;
    status?: number;
    message?: string | null;
    trace?: string | null;
};

export namespace StringObjectDictionaryListOperationResult {

    export enum outcome {
        SUCCESS = 'Success',
        FAILURE = 'Failure',
    }


}